+++ src/main/java/io/ticticboom/mods/mm/port/common/SlottedContainerScreen.java
@@ -1,6 +1,7 @@
 package io.ticticboom.mods.mm.port.common;
 
 import io.ticticboom.mods.mm.Ref;
+import io.ticticboom.mods.mm.client.gui.widgets.PortConfigPanel;
 import io.ticticboom.mods.mm.port.IPortBlockEntity;
 import io.ticticboom.mods.mm.port.IPortMenu;
 import net.minecraft.client.gui.GuiGraphics;
@@ -18,6 +19,7 @@ public class SlottedContainerScreen<T extends AbstractContainerMenu & IPortMenu>
     protected final T menu;
     protected final FormattedText header;
     protected ArrayList<Vec2> slots = new ArrayList<>();
+    protected final PortConfigPanel configPanel;
 
     public SlottedContainerScreen(T menu, Inventory inv, Component displayName) {
         super(menu, inv, displayName);
@@ -27,7 +29,27 @@ public class SlottedContainerScreen<T extends AbstractContainerMenu & IPortMenu>
         String name = menu.getModel().name();
         int subStrLength = Math.min(55, name.length());
         header = FormattedText.of(name.substring(0, subStrLength) + (subStrLength < 55 ? "" : "..."));
+        configPanel = new PortConfigPanel(menu.getBlockEntity());
         setupSlots();
+    }
+
+    @Override
+    protected void init() {
+        super.init();
+        configPanel.setPosition(this.leftPos + this.imageWidth + 2, this.topPos + 4);
+    }
+
+    @Override
+    public boolean mouseClicked(double mouseX, double mouseY, int button) {
+        if (configPanel.mouseClicked(mouseX, mouseY)) {
+            return true;
+        }
+        return super.mouseClicked(mouseX, mouseY, button);
+    }
+
+    @Override
+    protected boolean hasClickedOutside(double mouseX, double mouseY, int left, int top, int button) {
+        return super.hasClickedOutside(mouseX, mouseY, left, top, button) && !configPanel.isWithin(mouseX, mouseY);
     }
 
     private void setupSlots() {
@@ -66,6 +88,8 @@ public class SlottedContainerScreen<T extends AbstractContainerMenu & IPortMenu>
     public void render(GuiGraphics gfx, int mouseX, int mouseY, float partialTicks) {
         renderBackground(gfx);
         super.render(gfx, mouseX, mouseY, partialTicks);
+        configPanel.render(gfx, this.font, mouseX, mouseY);
         renderTooltip(gfx, mouseX, mouseY);
+        configPanel.renderTooltip(gfx, this.font, mouseX, mouseY);
     }
 }