+++ src/main/java/io/ticticboom/mods/mm/port/mekanism/infuse/MekanismInfusePortStorage.java
@@ -6,11 +6,16 @@ import io.ticticboom.mods.mm.port.common.INotifyChangeFunction;
 import io.ticticboom.mods.mm.port.mekanism.NotifyChangeContentsListener;
 import io.ticticboom.mods.mm.port.mekanism.chemical.MekanismChemicalPortStorage;
 import io.ticticboom.mods.mm.port.mekanism.chemical.MekanismChemicalPortStorageModel;
+import mekanism.api.MekanismAPI;
 import mekanism.api.chemical.ChemicalTankBuilder;
 import mekanism.api.chemical.IChemicalTank;
+import mekanism.api.chemical.infuse.IInfusionHandler;
 import mekanism.api.chemical.infuse.InfuseType;
 import mekanism.api.chemical.infuse.InfusionStack;
 import net.minecraftforge.common.capabilities.Capability;
+import net.minecraftforge.registries.IForgeRegistry;
+
+import java.util.function.Predicate;
 
 public class MekanismInfusePortStorage extends MekanismChemicalPortStorage<InfuseType, InfusionStack> {
 
@@ -19,8 +24,18 @@ public class MekanismInfusePortStorage extends MekanismChemicalPortStorage<Infus
     }
 
     @Override
-    protected IChemicalTank<InfuseType, InfusionStack> createTank(long capacity, INotifyChangeFunction changed) {
-        return ChemicalTankBuilder.INFUSION.createAllValid(capacity, new NotifyChangeContentsListener(changed));
+    protected IChemicalTank<InfuseType, InfusionStack> createTank(long capacity, Predicate<InfuseType> validator, INotifyChangeFunction changed) {
+        return ChemicalTankBuilder.INFUSION.create(capacity, validator, new NotifyChangeContentsListener(changed));
+    }
+
+    @Override
+    public Capability<IInfusionHandler> getChemicalCapability() {
+        return MekCapabilities.INFUSE;
+    }
+
+    @Override
+    protected IForgeRegistry<InfuseType> getChemicalRegistry() {
+        return MekanismAPI.infuseTypeRegistry();
     }
 
     @Override